+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.3
+
+- Added Visibilty Detector to run according to view port availability
+
+## 0.0.2
+
+- Added animations
+- And updated extensions for widget
+
 ## 0.0.1
 
 - Added Basic Widgets